backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-dotenv>=1.0.0
httpx>=0.26.0
ortools>=9.12
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
twilio>=8.13.0
python-multipart>=0.0.6
websockets>=12.0
